Убрал мягкое удаление из новостей статей и фиксированных статетей

+++ src/migrations/2019_08_19_135747_create_role_user_table.php
@@ -15,7 +15,7 @@ class CreateRoleUserTable extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->integer('role_id')->unsigned()->index();
             $table->timestamps();
         });
